+++ docs/fr_FR/Changelog.rst
@@ -1,6 +1,8 @@
 ChangeLog
 =========
 
+- Page EQ/avancé: Ajout bouton reset SW zigate (#2176).
+
 211202-BETA-1
 -------------